+++ trunk/nestbed/src/edu/clemson/cs/nestbed/common/management/profiling/NucleusManager.java
@@ -2,11 +2,11 @@
 /*
  * NucleusManager.java
  *
- * Network Embedded System Testbed (NESTBed)
+ * Network Embedded System Testbed (NESTbed)
  *
  * Copyright (C) 2006-2007
  * Dependable Systems Research Group
- * Department of Computer Science
+ * School of Computing
  * Clemson University
  * Andrew R. Dalton and Jason O. Hallstrom
  *
@@ -37,8 +37,8 @@ import edu.clemson.cs.nestbed.common.util.RemoteObserver;
 
 
 public interface NucleusManager extends Remote {
-    public int querySymbol(int    id,       String sourcePath,
-                           String moteType, String moteSerialID)
+    public long querySymbol(int    id,       String sourcePath,
+                            String moteType, String moteSerialID)
                                                     throws RemoteException;
 
     public boolean setSymbol(int    value,       int    id,
